Update project to Maven
