--- src/Launcher.java
@@ -1,5 +1,0 @@
-public class Launcher {
-    public static void main(String[] args){
-        Main.main(args);
-    }
-}
