+++ src/main/java/com/mastrHyperion98/FirstTimeController.java
@@ -1,5 +1,6 @@
-package scenes;
+package com.mastrHyperion98;
 
+import com.mastrHyperion98.struct.Controller;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
@@ -12,34 +13,30 @@ import javafx.stage.Stage;
 
 import java.io.IOException;
 
-public class LoginController {
-    private static struct.Controller myController;
+public class FirstTimeController {
+    private static Controller myController;
     private static Stage main_stage;
     @FXML
     private Text actiontarget;
     @FXML private PasswordField passwordField;
 
     @FXML protected void handleSubmitButtonAction(ActionEvent event) throws IOException {
-        boolean valid = myController.auth(passwordField.getText());
-
+        String password = passwordField.getText();
+        // validate password
+        boolean valid = myController.GenerateDatabase(password);
         if (valid){
-            FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("TableView.fxml"));
-            Parent root = fxmlLoader.load();
-            TableViewController tableViewController = fxmlLoader.<TableViewController>getController();
-            tableViewController.setController(myController);
+            Parent root = FXMLLoader.load(getClass().getResource("/scenes/login.fxml"));
+            LoginController.setController(myController);
+            LoginController.setStage(main_stage);
             Scene scene = new Scene(root, 400, 300);
             main_stage.setScene(scene);
-            main_stage.setMinWidth(1000);
-            main_stage.setMinHeight(400);
-            main_stage.setResizable(true);
         }
-
         else{
             actiontarget.setFill(Color.RED);
-            actiontarget.setText("ERROR: Password is Incorrect!");
+            actiontarget.setText("ERROR: Password is invalid!");
         }
     }
-    public static void setController(struct.Controller _controller){
+    public static void setController(Controller _controller){
         myController = _controller;
     }
