+++ src/main/resources/com/mastrHyperion98/EditEntry.fxml
@@ -1,0 +1,63 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+<?import javafx.scene.text.Text?>
+<AnchorPane minHeight="300.0" minWidth="400.0" xmlns="http://javafx.com/javafx/8.0.111" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.mastrHyperion98.UpdateEntryController">
+    <children>
+        <VBox alignment="CENTER" layoutX="131.0" layoutY="50.0" prefHeight="200.0" prefWidth="100.0" AnchorPane.bottomAnchor="5.0" AnchorPane.leftAnchor="5.0" AnchorPane.rightAnchor="5.0" AnchorPane.topAnchor="5.0">
+            <children>
+                <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Edit Entry"/>
+                <HBox alignment="CENTER" prefHeight="50.0" prefWidth="200.0" maxWidth="350.0" spacing="10.0">
+                    <children>
+                        <Label alignment="CENTER_LEFT" minWidth="100.0" text="Domain" />
+                        <TextField fx:id="tfDomain" HBox.hgrow="ALWAYS" />
+                    </children>
+                    <padding>
+                        <Insets right="30.0" />
+                    </padding>
+                </HBox>
+                <HBox alignment="CENTER" prefHeight="50.0" prefWidth="200.0" maxWidth="350.0" spacing="10.0">
+                    <children>
+                        <Label alignment="CENTER_LEFT" minWidth="100.0" text="Email" />
+                        <TextField fx:id="tfEmail" HBox.hgrow="ALWAYS" />
+                    </children>
+                    <padding>
+                        <Insets right="30.0" />
+                    </padding>
+                </HBox>
+                <HBox alignment="CENTER" prefHeight="50.0" prefWidth="200.0" maxWidth="350.0" spacing="10.0">
+                    <children>
+                        <Label alignment="CENTER_LEFT" minWidth="100.0" text="Username" />
+                        <TextField fx:id="tfUsername" HBox.hgrow="ALWAYS" />
+                    </children>
+                    <padding>
+                        <Insets right="30.0" />
+                    </padding>
+                </HBox>
+                <HBox alignment="CENTER" prefHeight="50.0" prefWidth="200.0" maxWidth="500.0" spacing="10.0">
+                    <children>
+                        <Label alignment="CENTER_RIGHT" minWidth="100.0" text="Password" />
+                        <TextField fx:id="tfPassword" HBox.hgrow="ALWAYS" />
+                        <Button mnemonicParsing="false" onAction="#btnGeneratePassowrd" text="Generate" />
+                    </children>
+                    <padding>
+                        <Insets right="30.0" />
+                    </padding>
+                </HBox>
+                <HBox alignment="CENTER_RIGHT">
+                    <children>
+                        <Button mnemonicParsing="false" onAction="#btnUpdateClicked" text="Update" />
+                    </children>
+                    <opaqueInsets>
+                        <Insets />
+                    </opaqueInsets>
+                    <padding>
+                        <Insets right="30.0" />
+                    </padding>
+                </HBox>
+            </children>
+        </VBox>
+    </children>
+</AnchorPane>
