+++ src/main/resources/com/mastrHyperion98/TableView.fxml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+<AnchorPane maxHeight="400.0" minHeight="400.0" minWidth="1000.0" xmlns="http://javafx.com/javafx/8.0.111" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.mastrHyperion98.TableViewController">
+    <children>
+        <VBox alignment="CENTER" layoutX="91.0" layoutY="85.0" spacing="10.0" AnchorPane.bottomAnchor="30.0" AnchorPane.leftAnchor="30.0" AnchorPane.rightAnchor="30.0" AnchorPane.topAnchor="30.0">
+            <children>
+                <TableView fx:id="data" prefHeight="300.0" prefWidth="400.0">
+                    <columns>
+                        <TableColumn fx:id="colId" prefWidth="40.0" minWidth="40.0" maxWidth="40.0" text="ID" />
+                        <TableColumn fx:id="colDomain" prefWidth="65.0" maxWidth="75.0" minWidth="65.0"
+                                     text="Domain" />
+                        <TableColumn fx:id="colEmail" prefWidth="75.0" text="Email" />
+                        <TableColumn fx:id="colUsername" prefWidth="75.0" text="Username" />
+                        <TableColumn fx:id="colPassword" prefWidth="100.0"  minWidth="100.00" text="Password" />
+                    </columns>
+                    <columnResizePolicy>
+                        <TableView fx:constant="CONSTRAINED_RESIZE_POLICY" />
+                    </columnResizePolicy>
+                </TableView>
+            </children>
+            <HBox alignment="CENTER" layoutX="91.0" layoutY="85.0" spacing="10.0" AnchorPane.bottomAnchor="30.0" AnchorPane.leftAnchor="30.0" AnchorPane.rightAnchor="30.0" AnchorPane.topAnchor="30.0">
+                <Button mnemonicParsing="false" onAction="#onOpenDialog" text="Add Entry" />
+                <Button mnemonicParsing="false" onAction="#onDeleteEntry" text="Delete Entry" />
+                <Button mnemonicParsing="false" onAction="#onUpdateDialog" text="Edit Entry" />
+            </HBox>
+        </VBox>
+    </children>
+</AnchorPane>
