+++ src/main/resources/com/mastrHyperion98/login.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.control.PasswordField?>
 <?import javafx.scene.layout.*?>
 <?import javafx.scene.text.Text?>
-<GridPane fx:controller="scenes.LoginController"
+<GridPane fx:controller="com.mastrHyperion98.LoginController"
           xmlns:fx="http://javafx.com/fxml" alignment="center" hgap="10" vgap="10">
     <padding><Insets top="25" right="25" bottom="10" left="25"/></padding>
     <Text text="Authentication"
